fix: Support calculating larger numbers

+++ src/common.rs
@@ -55,7 +55,7 @@ pub fn p(s: &str) -> String {
 
 // Adds commas to a string
 pub fn commas_from_string(n: &str) -> String {
-    let n = match n.parse::<f32>() {
+    let n = match n.parse::<f64>() {
         Ok(n) => n,
         Err(_) => 0.0,
     };
